// src/types/Wish.tsx
export type Wish = {
  id: number,
  title: string;
  description: string,
  dream_type: string;
  user_name: string;
  user_age: number;
  user_email: string;
  user_type: string;
  date: string;
  price: number;
  currency: string;
  attachment: string
  city: string;
  region: string;
  status: string;
  is_activated: false;
};